Refactored account screens and adapted logic
